+++ Changelog.rst
@@ -10,8 +10,6 @@ All notable changes to this project will be documented in this file.
 Added
 -----
 
-* Add release note automation workflow with Github actions.
-
 Changed
 -------
 
@@ -26,6 +24,26 @@ Removed
 
 Security
 --------
+
+`v0.6a1`_
+=========
+
+Added
+-----
+
+* Add release note automation workflow with Github actions.
+* COPY decompression method.(#61)
+
+Fixed
+-----
+
+* Fix retrieving Folder header information logics for codecs.(#62)
+
+
+Removed
+-------
+
+* Test symlink on windows.(#60)
 
 
 `v0.5b6`_
@@ -483,7 +501,8 @@ Changed
 
 
 .. History links
-.. _Unreleased: https://github.com/miurahr/py7zr/compare/v0.5b6...HEAD
+.. _Unreleased: https://github.com/miurahr/py7zr/compare/v0.6a1...HEAD
+.. _v0.6a1: https://github.com/miurahr/py7zr/compare/v0.5b6...v0.6a1
 .. _v0.5b6: https://github.com/miurahr/py7zr/compare/v0.5b5...v0.5b6
 .. _v0.5b5: https://github.com/miurahr/py7zr/compare/v0.5b4...v0.5b5
 .. _v0.5b4: https://github.com/miurahr/py7zr/compare/v0.5b3...v0.5b4